ka.tests: Fixes a floating conversion in functional tests.
VS2015 didn't accept constructing a `std::vector<float>` from double literal
values (e.g. `std::vector<float>{2.5, 3.8}` caused a compilation error).
Consequently, `float` is replaced by `double` as this does not impact test
semantics.

Reviewed-on: http://gerrit2.aldebaran.lan/1205
Tested-by: jmonnon <jmonnon@aldebaran.com>

+++ tests/ka/test_functional.cpp
@@ -526,7 +526,7 @@ TEST(FunctionalComposeT, Bind) {
   ASSERT_TRUE(str_even("20"));
   ASSERT_FALSE(str_even("21"));
 
-  using V = vector<float>;
+  using V = vector<double>;
   auto len = [](V const& x) -> int { return x.size(); };
   auto len_even = comp_even(len); // len_even: vector<float> → int → bool
   ASSERT_TRUE(len_even(V{1.4, 76.2}));
